Review: extract config on higher level
