+++ cmd/agent/main.go
@@ -2,8 +2,8 @@ package main
 
 import (
 	"fmt"
+	"github.com/PostScripton/go-metrics-and-alerting-collection/config"
 	"github.com/PostScripton/go-metrics-and-alerting-collection/internal/agent"
-	"github.com/PostScripton/go-metrics-and-alerting-collection/internal/agent/config"
 	"github.com/PostScripton/go-metrics-and-alerting-collection/internal/client"
 	"github.com/PostScripton/go-metrics-and-alerting-collection/internal/monitoring"
 	"github.com/PostScripton/go-metrics-and-alerting-collection/internal/repository/memory"
@@ -17,8 +17,7 @@ func main() {
 	zerolog.TimeFieldFormat = zerolog.TimeFormatUnix
 	log.Logger = log.Output(zerolog.ConsoleWriter{Out: os.Stderr, TimeFormat: "03:04:05PM"})
 
-	cfg := config.NewConfig()
-	log.Info().Interface("config", cfg).Send()
+	cfg := config.NewAgentConfig()
 
 	baseURI := fmt.Sprintf("http://%s", cfg.Address)
