+++ config/config.go
@@ -1,0 +1,9 @@
+package config
+
+type CommonConfig struct {
+	Address string `env:"ADDRESS"`
+	Key     string `env:"KEY"`
+}
+
+const defaultAddress = "localhost:8080"
+const defaultKey = ""
